+++ src/main/java/it/polimi/ingsw/view/client/gui/StartTurn.java
@@ -48,7 +48,7 @@ public class StartTurn extends JDesktopPane{
         JLabel label = ImageHandler.setImage("resources/Graphics/Texts/is_starting_the_turn.png", 99, 99, frameSize.width * 85/100, frameSize.height * 22/100);
         JLabel label2 = ImageHandler.setImage("resources/Graphics/Texts/it's_your_turn.png", 99, 99, frameSize.width * 60/100, frameSize.height * 22/100);
         label.setBounds((int) (frameSize.width * 7.5/100), (int) (frameSize.height * 39.5/100), frameSize.width * 85/100, frameSize.height * 22/100);
-        label2.setBounds((int) (frameSize.width * 20/100), frameSize.height * 39/100, frameSize.width * 60/100, frameSize.height * 22/100);
+        label2.setBounds((frameSize.width * 20/100), frameSize.height * 39/100, frameSize.width * 60/100, frameSize.height * 22/100);
         JLabel otherName = new JLabel(nameChoosing);
         otherName.setBounds((int) (((double)frameSize.width * 48/100) - ((otherName.getText().length() * bold) / 2)), (int) (frameSize.height * 34.5/100), frameSize.width * 60/100, frameSize.width * 5/100);
         otherName.setFont(felixBold);
@@ -63,7 +63,7 @@ public class StartTurn extends JDesktopPane{
 
 
         close.addActionListener(new Close());
-        close.setBounds((int) ((frameSize.width * 50/100) - (buttonSize.width / 2)), (int) (frameSize.height * 81 / 100),  buttonSize.width, buttonSize.height);
+        close.setBounds(((frameSize.width * 50/100) - (buttonSize.width / 2)), (frameSize.height * 81 / 100),  buttonSize.width, buttonSize.height);
         add(close);
